Moved most of the runtime stuff out of our root "index.ts"
