+++ bindings/web/src/Runtime.ts
@@ -1,7 +1,31 @@
 import Shape from "./Shape";
-import { Capability, Imports, Model, Output } from ".";
-import { TensorFlowLiteMimeType } from "./builtin";
-import { IdentityInputs } from "@tensorflow/tfjs-core";
+
+/**
+ * Something which consumes outputs generated by the Rune.
+ */
+export interface Output {
+    consume(data: Uint8Array): void;
+}
+
+/**
+ * Inputs provided by the application.
+  */
+export interface Capability {
+    generate(dest: Uint8Array): void;
+    setParameter(name: string, value: number): void;
+}
+
+export interface Imports {
+    createOutput(type: number): Output;
+    createCapability(type: number): Capability;
+    createModel(mimetype: string, model: ArrayBuffer): Promise<Model>;
+    log(message: string): void;
+}
+
+export interface Model {
+    transform(inputArray: Uint8Array[], inputDimensions: Shape[], outputArray: Uint8Array[], outputDimensions: Shape[]): void;
+}
+
 
 type CapabilityInfo = {
     capabilityType: number,
