+++ bindings/web/src/index.ts
@@ -1,37 +1,21 @@
-export { Runtime } from "./Runtime";
-export { CapabilityType, Builder } from "./facade";
-
-import { Tensor } from "@tensorflow/tfjs-core";
-import { Builder, InputDescription } from "./facade";
+import { Builder } from "./facade";
 import { TensorFlowModel, TensorFlowLiteMimeType } from "./builtin";
-import Shape from "./Shape";
 
-/// Something which consumes outputs generated by the Rune.
-export interface Output {
-    consume(data: Uint8Array): void;
-}
+export const Capabilities = {
+    "rand": 1,
+    "sound": 2,
+    "accel": 3,
+    "image": 4,
+    "raw": 5,
+} as const;
 
-/**
- * Inputs provided by the application.
-  */
-export interface Capability {
-    generate(dest: Uint8Array): void;
-    setParameter(name: string, value: number): void;
-}
-
-export interface Imports {
-    createOutput(type: number): Output;
-    createCapability(type: number): Capability;
-    createModel(mimetype: string, model: ArrayBuffer): Promise<Model>;
-    log(message: string): void;
-}
-
-export interface Model {
-    transform(inputArray: Uint8Array[], inputDimensions: Shape[], outputArray: Uint8Array[], outputDimensions: Shape[]): void;
-}
+export type CapabilityType = keyof typeof Capabilities;
 
 /**
  * Use a high level builder API to initialize the Rune runtime.
+ *
+ * Check out the "Runtime" module if you need tighter control over the runtime
+ * or want to avoidunnecessary indirection/copies.
  */
 export function builder(): Builder {
     return new Builder()
